requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.7
gunicorn==20.1.0
pytz==2021.1
sqlparse==0.4.2
psycopg2-binary==2.9
whitenoise==5.2.0
djangorestframework==3.12.4
python-dotenv==0.14.0
django-crontab==0.7.1
